requirements.txt
requests==2.31.0
websockets==12.0
pyyaml==6.0.1
pydantic==2.6.0
pandas==2.2.0
python-dotenv==1.0.1
asyncio
fastapi
uvicorn
httpx
